chatbot/requirements.txt
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.0
langchain-huggingface>=0.0.3
pydantic==2.11.5
requests==2.32.4
streamlit==1.45.1
langchain==0.3.25
langchain-core==0.3.65
langchain-community==0.3.25
langchain-huggingface==0.2.0
faiss-cpu==1.11.0
numpy==2.3.0
unstructured==0.17.2
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.1
fastapi
uvicorn
python-dotenv